Steps 11 + 12 — admin/page.tsx becomes a 70-line router; shared markdown renderer.
Step 12 (client/src/components/comments/markdown.ts)
- Both <CommentItem> (renderMarkdown) and <CommentForm> (renderPreview)
  re-declared the marked + DOMPurify + linkifyMentions pipeline with
  subtly different ALLOWED_TAGS lists — preview was MISSING sup, sub,
  table, thead, tbody, tr, th, td that the actual render allowed, so
  the preview silently lied about what would appear after submit.
- Extract renderCommentMarkdown() as the single source of truth.
  Browser-only `window` guard moved into the renderer; SSR returns ''
  without trying to call DOMPurify.
- CommentForm + CommentItem both `import { renderCommentMarkdown }`
  and drop the local helpers.

Step 11 (client/src/components/admin/ + client/src/app/admin/page.tsx)
- page.tsx was 1335 lines with 7 inline tab components, shared
  styled-components, an inline reputation-flavour helper, an inline
  formatDate, and an IIFE for the Top Bookmarked rank card.
- New layout:
    components/admin/_shared/styles.tsx  — StatGrid/StatCard/AdminBadge/
                                            BarChart/Bar/RankList/etc.
    components/admin/_shared/formatDate.ts
    components/admin/OverviewTab.tsx
    components/admin/CommentsTab.tsx     — owns reputationFlavour +
                                            tab-local InlineActions/
                                            MiniButton/Sparkline
    components/admin/UsersTab.tsx        — owns groupColour helper + the
                                            FILTER_GROUPS tuple replaces
                                            the bespoke groupCounts object
    components/admin/AuditTab.tsx
    components/admin/AnalyticsTab.tsx    — three near-identical Top Posts
                                            cards collapsed into a single
                                            TopPostsRankCard generic
                                            component; the bolted-on IIFE
                                            for Top Bookmarked is gone
    components/admin/NotificationsTab.tsx
    components/admin/ReportsTab.tsx      — STATUS_FILTERS tuple + named
                                            statusColour helper replace the
                                            inline ternary cascade
- page.tsx is now a 70-line router driven by a TABS tuple — adding a
  tab is two lines (tuple entry + the conditional render).
- Local AdminReport interface deleted (now in @derecksnotes/shared from
  step 1).
- Inline styled `Badge` renamed to `AdminBadge` to avoid clashing with
  the existing PageStyles convention; same shape.

Tested: bunx tsc --noEmit clean on server + client; bun test 45 pass.
