+++ client/src/components/admin/_shared/formatDate.ts
@@ -1,0 +1,9 @@
+export function formatAdminDate(iso: string): string {
+  return new Date(iso).toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit'
+  });
+}
